move to kotlin
make depth an int rather than byte

+++ src/main/java/net/ropes4k/impl/ReverseRope.java
@@ -38,7 +38,7 @@ public final class ReverseRope extends AbstractRope {
     }
 
     @Override
-    public byte depth() {
+    public int depth() {
         return RopeUtilities.Companion.depth(rope);
     }
